deps: no default features for tempo crates (#662)
Even though a bunch of deps in all tempo crates were marked as optional,
they were also part of the `default` feature, which means all of them
were being added anyway for downstream consumers. As an example,
depending on `tempo-alloy` at the moment compiles `reth-cli-commands`
and everything downstream (most of Reth).

Adding `default-features = false` to the workspace for the tempo crates
fixes this and depending on `tempo-alloy` now only compiles two reth
crates related to primitives, which is not that bad.

---------

Co-authored-by: Matthias Seitz <matthias.seitz@outlook.de>
Co-authored-by: Arsenii Kulikov <klkvrr@gmail.com>