feat: add gas schedule for access keys  (#1413)
Currently setting an access key using key authorization and setting
spending limits is basically free in the protocol.
Since every authorized key and every spending limit requires 1 SSTORE,
this makes the node potentially vuln to DOS.

In this PR, we update the spec, and add extra costs for key
authorization. The rationale, and exact gas changes can be seen in the
docs.

We also add better `eth_estimateGas` support for the access key type.

---------

Co-authored-by: 0xKitsune <77890308+0xKitsune@users.noreply.github.com>
Co-authored-by: Jennifer <jenpaff0@gmail.com>
Co-authored-by: 0xKitsune <0xkitsune@protonmail.com>
Co-authored-by: Arsenii Kulikov <klkvrr@gmail.com>